Check type in Variable.setVal(); move step/* to steps/*; move evalWithContext to Context.eval()
