+++ src/steps/goto-step.js
@@ -1,5 +1,5 @@
 const $ = require('jquery');
-const Context = require('../context').Context;
+const Context = require('../context');
 const BaoStep = require('./base-step');
 
 /** A step that contains goto clause. */
