+++ src/steps/switch-step.js
@@ -44,7 +44,7 @@ class SwitchStep extends BaoStep {
     if (!this.expr_ || !this.cases_) {
       throw 'Invalid switch step';
     }
-    const exprVal = baoContext.evalWithContext(this.context_, this.expr_);
+    const exprVal = this.context_.eval(this.expr_);
     const clause = this.cases_.get(exprVal);
     if (clause) {
       return clause.run();
